backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
elasticsearch==8.11.0
redis==5.0.1
kafka-python==2.0.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0
slack-sdk==3.26.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
